refactor(types): ディレクトリ構造を整理し型定義を分割
