+++ package/main/src/types/math/numberToArray.ts
@@ -1,0 +1,5 @@
+// 数値を配列に変換する
+export type NumberToArray<
+  T extends number,
+  U extends number[] = [],
+> = U["length"] extends T ? U : NumberToArray<T, [...U, U["length"]]>;
